fixed game not running
issue was that some sprites were being looked for in the local pictures folder so i copied the project's bullet sprite and renamed to each missing sprite, then rebuilt the MGCB thing.

also, in fontMain.spritefont, SPACEFONT was not found so i changed that to Arial.

also, in the XMLParser, the XML files were also being looked for in the local path so i changed that to go from the current directory.